Create the class for the run function
